kinematics: build the beam charge array as float64
particle 1.0.0 changed PDGID.charge from float to fractions.Fraction:

    particle 0.26.2:  PDGID(211).charge -> 1.0            (float)
    particle 1.0.0 :  PDGID(211).charge -> Fraction(1, 1)

np.array([p1.charge, p2.charge]) therefore falls back to dtype=object.
Numpy has no object-dtype loop for isfinite, so any np.isclose/np.allclose
over the beam record raises

    TypeError: ufunc 'isfinite' not supported for the input types

_repair_initial_beam concatenates this beam record onto the event arrays, so
the object dtype propagates into event.charge for every model with a
prepended beam, not just the beam slots.

This currently fails tests/test_common.py::test_models_beam for all 23
non-xfailed models (SIBYLL, QGSJET, Pythia, Phojet, Sophia, UrQMD, DPMJET,
EPOS). The values themselves are correct and equal on both sides; only the
dtype is wrong. CI has not caught it because chromo declares particle with
no upper bound and older resolutions returned floats.

Pinning the charge dtype fixes it at the source and keeps the array usable
for downstream arithmetic. The adjacent fields already pin their dtypes.

+++ src/chromo/kinematics.py
@@ -165,7 +165,7 @@ class EventKinematicsBase:
         self._beam_data = {
             "pid": np.array([int(self.p1), int(self.p2)]),
             "status": np.array([4, 4]),
-            "charge": np.array([self.p1.charge, self.p2.charge]),
+            "charge": np.array([self.p1.charge, self.p2.charge], dtype=np.float64),
             "px": np.zeros((2,), dtype=np.float64),
             "py": np.zeros((2,), dtype=np.float64),
             "pz": event_like.pz,
